+++ core/src/main/kotlin/org/sert2521/sertain/motors/MotorController.kt
@@ -1,6 +1,5 @@
 package org.sert2521.sertain.motors
 
-import com.ctre.phoenix.motorcontrol.can.BaseMotorController
 import org.sert2521.sertain.units.*
 import com.ctre.phoenix.motorcontrol.NeutralMode as CtreNeutralMode
 
@@ -40,7 +39,7 @@ abstract class MotorController {
 
 fun<T: MotorId> motorController(id: T, vararg followerIds: MotorId, configure: MotorController.() -> Unit = {}): MotorController {
     return when (id) {
-        is TalonId  -> TalonMotorController(id,  *followerIds, configure = configure)
+        is TalonId  -> RealTalonMotorController(id,  *followerIds, configure = configure)
         is VictorId -> VictorMotorController(id, *followerIds, configure = configure)
         else        -> throw NotImplementedError("A Motor Controller is not implemented for this ID")
     }
